requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.5
django-cors-headers==2.1.0
djangorestframework==3.7.0
idna==2.6
pkg-resources==0.0.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
websockets==4.0.1
